+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Release Versions:
 ## Upcoming changes (in development)
 
 - Fix test actions and workflows (#147)
+- Use rclpy lifecycle node as base class for lifecycle components (#134, #140, #146)
 
 ## 2.1.1
 ### October 28, 2022